+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -292,7 +292,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-08T03:27:03.024Z",
+    "2025-10-08T07:01:29.944Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -332,7 +332,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-08T03:27:03.024Z"))
+  updated: z.date().default(() => new Date("2025-10-08T07:01:29.944Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {