+++ packages/orq-rc/RELEASES.md
@@ -4149,3 +4149,13 @@ Based on:
 - [typescript v3.14.0-rc.15] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.15 - packages/orq-rc
+
+## 2025-10-08 07:03:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.0 (2.723.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.16 - packages/orq-rc